Tune config usage

+++ src/zoomer_navigation.c
@@ -6,10 +6,10 @@ Vector world (const Camera camera, const Vector pos) {
 }
 
 void camera_update(Camera *camera, const Config config, const float dt, const Mouse mouse, const Vector window_size) {
-    if (fabs(camera->delta_scale) > 0.5) {
-        Vector p0 = VectorScale(VectorSubtract(camera->scale_pivot, VectorScale(window_size, 0.5f)), 1 / camera->scale);
+    if (fabsf(camera->delta_scale) > 0.5f) {
+        Vector p0 = VectorScale(VectorSubtract(camera->scale_pivot, VectorScale(window_size, 0.5f)), 1.0f / camera->scale);
         camera->scale = fmaxf(camera->scale + camera->delta_scale * dt, config.min_scale);
-        Vector p1 = VectorScale(VectorSubtract(camera->scale_pivot, VectorScale(window_size, 0.5f)), 1 / camera->scale);
+        Vector p1 = VectorScale(VectorSubtract(camera->scale_pivot, VectorScale(window_size, 0.5f)), 1.0f / camera->scale);
         VectorAddMut(&camera->position, VectorSubtract(p0, p1));
 
         camera->delta_scale -= camera->delta_scale * dt * config.scale_friction;
